Basic Skeleton of sign up page, moved log in page to its own, sign up as its own, index is now empty, have to figure out the navigation
